// src/shared/interfaces/old-database-users.interface.ts
export interface OldUser {
  user_id?: string;
  user_login?: string;
  user_email?: string;
  display_name?: string;
  first_name?: string;
  last_name?: string;
  description?: string;
  dni_number?: string;
  phone_number?: string;
  localidad?: string;
  city?: string;
  empleo?: string;
  account_status?: string;
  nsocio?: string;
  user_registered?: string;
  roles?: string[];
}